status online/offline

+++ lib/pages/chat.dart
@@ -450,7 +450,15 @@ class _ChatPageState extends State<ChatPage> {
   }
 
   Future<void> _logout(BuildContext context) async {
-    // Hapus status login dari SharedPreferences
+    final user = FirebaseAuth.instance.currentUser;
+    if (user == null) return;
+    String status = 'Offline';
+    FirebaseDatabase.instance.ref("users/${user.uid}").update({
+      "status": status,
+      "lastSeen": DateTime.now().millisecondsSinceEpoch,
+    });
+
+// Hapus status login dari SharedPreferences
     SharedPreferences prefs = await SharedPreferences.getInstance();
     await prefs.clear();
