+++ lib/pages/lesson.dart
@@ -221,6 +221,14 @@ class _LessonPageState extends State<LessonPage> {
   }
 
   Future<void> _logout(BuildContext context) async {
+    final user = FirebaseAuth.instance.currentUser;
+    if (user == null) return;
+    String status = 'Offline';
+    FirebaseDatabase.instance.ref("users/${user.uid}").update({
+      "status": status,
+      "lastSeen": DateTime.now().millisecondsSinceEpoch,
+    });
+
     // Hapus status login dari SharedPreferences
     SharedPreferences prefs = await SharedPreferences.getInstance();
     await prefs.clear();
@@ -906,13 +914,6 @@ class _LessonPageState extends State<LessonPage> {
                                         'data': _data![key],
                                         'user': _isAdmin,
                                       });
-                                      // if (_isAdmin) {
-                                      //   _navigateAndRefresh('/datalog', {
-                                      //     'uid': key,
-                                      //     'matkul': _matkul,
-                                      //     'type': _data![key]['type'],
-                                      //   });
-                                      // }
                                     },
                                   ),
                                 ),
